% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatOps.R
\name{MatMult}
\alias{MatMult}
\title{MatMult}
\usage{
MatMult(A, B)
}
\arguments{
\item{A}{First numeric matrix}

\item{B}{Second numeric matrix}
}
\value{
a numeric matrix
}
\description{
Multiply two numeric matrices
}
\examples{
C <- MatMult(A, B)
}
