% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatOps.R
\name{MatTCross}
\alias{MatTCross}
\title{MatTCross}
\usage{
MatTCross(A, B)
}
\arguments{
\item{A}{First numeric matrix}

\item{B}{Second numeric matrix}
}
\value{
a numeric matrix
}
\description{
Calculate a matrix transpose cross-product
}
\examples{
C <- MatMult(A, B)
}
